requirements.txt
Adafruit-Blinka==8.39.2
adafruit-circuitpython-ads1x15==2.2.25
adafruit-circuitpython-busdevice==5.2.9
adafruit-circuitpython-connectionmanager==3.1.0
adafruit-circuitpython-dht==4.0.4
adafruit-circuitpython-requests==4.0.0
adafruit-circuitpython-typing==1.10.3
Adafruit-PlatformDetect==3.63.0
Adafruit-PureIO==1.1.11
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
lz4==4.3.3
psutil==5.9.8
pyftdi==0.55.4
pyserial==3.5
python-decouple==3.8
pyusb==1.2.1
requests==2.32.2
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
SqliteCloud==0.0.77
sysv-ipc==1.1.0
typing_extensions==4.12.0
urllib3==2.2.1
websockets==12.0
